WIP: Two versions of the aap .cpu files
Create two versions of the aap cpu files, one containing only a couple
of instructions (for debugging) and one containing more instructions.

These files are written in a manor that I think is correct, but doesn't
actually work; we need to fix cgen itself in order for these files to
work.

In order to actually make some aap progress, I'm going to rework these
files inline with the "invalid" way that other targets are implemented.